+++ packages/devtools_app/test/deep_link_vlidation/deep_links_screen_test.dart
@@ -45,6 +45,16 @@ final domainErrorlinkData = LinkData(
   domainErrors: [DomainError.existence],
 );
 
+final pathErrorlinkData = LinkData(
+  domain: 'www.google.com',
+  path: '/abcd',
+  os: [PlatformOS.android, PlatformOS.ios],
+  pathErrors: {
+    PathError.intentFilterActionView,
+    PathError.intentFilterDefault,
+  },
+);
+
 void main() {
   setUp(() {
     setGlobal(ServiceConnectionManager, FakeServiceConnectionManager());
@@ -174,7 +184,7 @@ void main() {
     );
 
     testWidgetsWithWindowSize(
-      'shows notification cards when there are errors',
+      'shows notification cards when there are domain errors',
       windowSize,
       (WidgetTester tester) async {
         final deepLinksController = DeepLinksTestController();
@@ -183,6 +193,26 @@ void main() {
             FlutterProject(path: '/abc', androidVariants: ['debug', 'release']);
 
         deepLinksController.allValidatedLinkDatas = [domainErrorlinkData];
+        await pumpDeepLinkScreen(
+          tester,
+          controller: deepLinksController,
+        );
+
+        expect(find.byType(DeepLinkPage), findsOneWidget);
+        expect(find.byType(DeepLinkListView), findsOneWidget);
+        expect(find.byType(NotificationCard), findsOneWidget);
+      },
+    );
+    testWidgetsWithWindowSize(
+      'shows notification cards when there are path errors',
+      windowSize,
+      (WidgetTester tester) async {
+        final deepLinksController = DeepLinksTestController();
+
+        deepLinksController.selectedProject.value =
+            FlutterProject(path: '/abc', androidVariants: ['debug', 'release']);
+
+        deepLinksController.allValidatedLinkDatas = [pathErrorlinkData];
         await pumpDeepLinkScreen(
           tester,
           controller: deepLinksController,
@@ -345,7 +375,7 @@ void main() {
             domain: 'www.domain2.com',
             path: '/path',
             os: [PlatformOS.ios],
-            pathError: true,
+            pathErrors: {PathError.intentFilterActionView},
           ),
           LinkData(
             domain: 'www.google.com',
@@ -469,6 +499,76 @@ void main() {
         expect(widgetBCenter.dy < widgetCCenter.dy, true);
       },
     );
+
+    testWidgetsWithWindowSize(
+      'path view',
+      windowSize,
+      (WidgetTester tester) async {
+        final deepLinksController = DeepLinksTestController();
+
+        deepLinksController.selectedProject.value =
+            FlutterProject(path: '/abc', androidVariants: ['debug', 'release']);
+
+        final linkDatas = [
+          LinkData(
+            domain: 'www.domain1.com',
+            path: '/path1',
+            os: [PlatformOS.android],
+            domainErrors: [DomainError.existence],
+          ),
+          LinkData(
+            domain: 'www.domain2.com',
+            path: '/path2',
+            os: [PlatformOS.ios],
+            pathErrors: {PathError.intentFilterActionView},
+          ),
+          LinkData(
+            domain: 'www.google.com',
+            path: '/path3',
+            os: [PlatformOS.android, PlatformOS.ios],
+          ),
+        ];
+
+        deepLinksController.allValidatedLinkDatas = linkDatas;
+
+        await pumpDeepLinkScreen(
+          tester,
+          controller: deepLinksController,
+        );
+
+        await tester.tap(find.text('Path view'));
+        await tester.pumpAndSettle(const Duration(milliseconds: 500));
+
+        expect(find.text('/path1'), findsOneWidget);
+        expect(find.text('/path2'), findsOneWidget);
+        expect(find.text('/path3'), findsOneWidget);
+
+        // Only show links with path error.
+        deepLinksController.updateDisplayOptions(
+          removedFilter: FilterOption.noIssue,
+        );
+
+        await tester.pumpAndSettle();
+
+        expect(find.text('/path1'), findsNothing);
+        expect(find.text('/path2'), findsOneWidget);
+        expect(find.text('/path3'), findsNothing);
+
+        // Only show links with no issue.
+        deepLinksController.updateDisplayOptions(
+          removedFilter: FilterOption.failedPathCheck,
+        );
+        deepLinksController.updateDisplayOptions(
+          addedFilter: FilterOption.noIssue,
+        );
+
+        await tester.pumpAndSettle();
+
+        expect(find.text('/path1'), findsOneWidget);
+        expect(find.text('/path2'), findsNothing);
+        expect(find.text('/path3'), findsOneWidget);
+      },
+    );
   });
 }
 
@@ -482,8 +582,9 @@ class DeepLinksTestController extends DeepLinksController {
       domainErrorCount: getLinkDatasByDomain
           .where((element) => element.domainErrors.isNotEmpty)
           .length,
-      pathErrorCount:
-          getLinkDatasByPath.where((element) => element.pathError).length,
+      pathErrorCount: getLinkDatasByPath
+          .where((element) => element.pathErrors.isNotEmpty)
+          .length,
     );
     pagePhase.value = PagePhase.linksValidated;
   }